+++ infoblox_client/objects.py
@@ -176,7 +176,11 @@ class InfobloxObject(BaseObject):
 
     _fields - fields that represents NIOS object (WAPI fields) and
         are sent to NIOS on object creation
-    _search_fields - fields that can be used to find object on NIOS side
+    _search_for_update_fields - field/fields used to find an object during an
+        update operation. this should be the smallest number of fields that
+        uniquely identify an object
+    _all_searchable_fields - all fields that can be used to find object on NIOS
+        side
     _updateable_search_fields - fields that can be used to find object on
         NIOS side, but also can be changed, so has to be sent on update.
     _shadow_fields - fields that object usually has but they should not
@@ -202,7 +206,8 @@ class InfobloxObject(BaseObject):
         no versioned class lookup needed.
     """
     _fields = []
-    _search_fields = []
+    _search_for_update_fields = []
+    _all_searchable_fields = []
     _updateable_search_fields = []
     _shadow_fields = []
     _infoblox_type = None
@@ -261,14 +266,16 @@ class InfobloxObject(BaseObject):
     def to_dict(self, search_fields=None):
         """Builds dict without None object fields"""
         fields = self._fields
-        if search_fields == 'only':
-            fields = self._search_fields
+        if search_fields == 'update':
+            fields = self._search_for_update_fields
+        elif search_fields == 'all':
+            fields = self._all_searchable_fields
         elif search_fields == 'exclude':
             # exclude search fields for update actions,
             # but include updateable_search_fields
             fields = [field for field in self._fields
                       if field in self._updateable_search_fields or
-                      field not in self._search_fields]
+                      field not in self._search_for_update_fields]
 
         return {field: self.field_to_dict(field) for field in fields
                 if getattr(self, field, None) is not None}
@@ -286,8 +293,12 @@ class InfobloxObject(BaseObject):
         return parse_class.from_dict(connector, return_dict)
 
     @classmethod
-    def create(cls, connector, check_if_exists=True,
-               update_if_exists=False, **kwargs):
+    def create_check_exists(cls, connector, check_if_exists=True,
+                            update_if_exists=False, **kwargs):
+        # obj_created is used to check if object is being created or
+        # pre-exists. obj_created is True if object is not pre-exists
+        # and getting created with this function call
+        obj_created = False
         local_obj = cls(connector, **kwargs)
         if check_if_exists:
             if local_obj.fetch(only_ref=True):
@@ -296,12 +307,13 @@ class InfobloxObject(BaseObject):
                          {'obj_type': local_obj.infoblox_type,
                           'ib_obj': local_obj})
                 if not update_if_exists:
-                    return local_obj
+                    return local_obj, obj_created
         reply = None
         if not local_obj.ref:
             reply = connector.create_object(local_obj.infoblox_type,
                                             local_obj.to_dict(),
                                             local_obj.return_fields)
+            obj_created = True
             LOG.info("Infoblox %(obj_type)s was created: %(ib_obj)s",
                      {'obj_type': local_obj.infoblox_type,
                       'ib_obj': local_obj})
@@ -311,14 +323,24 @@ class InfobloxObject(BaseObject):
                                             update_fields,
                                             local_obj.return_fields)
             LOG.info('Infoblox object was updated: %s', local_obj.ref)
-        return cls._object_from_reply(local_obj, connector, reply)
+        return cls._object_from_reply(local_obj, connector, reply), obj_created
+
+    @classmethod
+    def create(cls, connector, check_if_exists=True,
+               update_if_exists=False, **kwargs):
+        ib_object, _ = (
+            cls.create_check_exists(connector,
+                                    check_if_exists=check_if_exists,
+                                    update_if_exists=update_if_exists,
+                                    **kwargs))
+        return ib_object
 
     @classmethod
     def _search(cls, connector, return_fields=None,
                 search_extattrs=None, force_proxy=False,
                 max_results=None, **kwargs):
         ib_obj_for_search = cls(connector, **kwargs)
-        search_dict = ib_obj_for_search.to_dict(search_fields='only')
+        search_dict = ib_obj_for_search.to_dict(search_fields='all')
         if return_fields is None and ib_obj_for_search.return_fields:
             return_fields = ib_obj_for_search.return_fields
         # allow search_extattrs to be instance of EA class
@@ -363,7 +385,7 @@ class InfobloxObject(BaseObject):
                 self.update_from_dict(reply)
                 return True
 
-        search_dict = self.to_dict(search_fields='only')
+        search_dict = self.to_dict(search_fields='update')
         return_fields = [] if only_ref else self.return_fields
         reply = self.connector.get_object(self.infoblox_type,
                                           search_dict,
@@ -426,11 +448,12 @@ class InfobloxObject(BaseObject):
 
 class Network(InfobloxObject):
     _fields = ['network_view', 'network', 'template',
-               'options', 'members', 'extattrs']
-    _search_fields = ['network_view', 'network']
+               'options', 'members', 'extattrs', 'comment']
+    _search_for_update_fields = ['network_view', 'network']
+    _all_searchable_fields = _search_for_update_fields
     _shadow_fields = ['_ref']
     _return_fields = ['network_view', 'network', 'options', 'members',
-                      'extattrs']
+                      'extattrs', 'comment']
     _remap = {'cidr': 'network'}
 
     @classmethod
@@ -531,8 +554,11 @@ class HostRecord(InfobloxObject):
 class HostRecordV4(HostRecord):
     """HostRecord for IPv4"""
     _fields = ['ipv4addrs', 'view', 'extattrs', 'name', 'zone',
-               'configure_for_dns']
-    _search_fields = ['view', 'ipv4addr', 'name', 'zone']
+               'configure_for_dns', 'network_view', 'mac', 'ttl',
+               'comment']
+    _search_for_update_fields = ['view', 'ipv4addr', 'name',
+                                 'zone', 'network_view', 'mac']
+    _all_searchable_fields = _search_for_update_fields
     _updateable_search_fields = ['name']
     _shadow_fields = ['_ref', 'ipv4addr']
     _return_fields = ['ipv4addrs', 'extattrs']
@@ -564,8 +590,10 @@ class HostRecordV4(HostRecord):
 class HostRecordV6(HostRecord):
     """HostRecord for IPv6"""
     _fields = ['ipv6addrs', 'view', 'extattrs',  'name', 'zone',
-               'configure_for_dns']
-    _search_fields = ['ipv6addr', 'view', 'name', 'zone']
+               'configure_for_dns', 'network_view', 'ttl', 'comment']
+    _search_for_update_fields = ['ipv6addr', 'view', 'name',
+                                 'zone', 'network_view']
+    _all_searchable_fields = _search_for_update_fields
     _updateable_search_fields = ['name']
     _shadow_fields = ['_ref', 'ipv6addr']
     _return_fields = ['ipv6addrs', 'extattrs']
@@ -593,6 +621,16 @@ class HostRecordV6(HostRecord):
         return [IPv6.from_dict(ip_addr) for ip_addr in ips_v6]
 
     _custom_field_processing = {'ipv6addrs': _build_ipv6.__func__}
+
+
+class IPv6HostAddress(InfobloxObject):
+    _infoblox_type = 'record:host_ipv6addr'
+    _fields = ['duid', 'network_view', 'host']
+    _search_for_update_fields = ['duid', 'network_view']
+    _all_searchable_fields = _search_for_update_fields
+    _shadow_fields = ['_ref']
+    _return_fields = ['host']
+    _ip_version = 6
 
 
 class SubObjects(BaseObject):
@@ -695,7 +733,9 @@ class IPRange(InfobloxObject):
     _fields = ['start_addr', 'end_addr', 'network_view',
                'network', 'extattrs', 'disable']
     _remap = {'cidr': 'network'}
-    _search_fields = ['network_view', 'start_addr', 'network']
+    _search_for_update_fields = ['network_view', 'start_addr',
+                                 'end_addr', 'network']
+    _all_searchable_fields = _search_for_update_fields
     _shadow_fields = ['_ref']
     _return_fields = ['start_addr', 'end_addr', 'network_view', 'extattrs']
 
@@ -740,8 +780,9 @@ class FixedAddress(InfobloxObject):
 class FixedAddressV4(FixedAddress):
     _infoblox_type = 'fixedaddress'
     _fields = ['ipv4addr', 'mac', 'network_view', 'extattrs', 'network',
-               'options']
-    _search_fields = ['ipv4addr', 'mac', 'network_view', 'network']
+               'options', 'comment']
+    _search_for_update_fields = ['ipv4addr', 'mac', 'network_view', 'network']
+    _all_searchable_fields = _search_for_update_fields
     _shadow_fields = ['_ref', 'ip']
     _return_fields = ['ipv4addr', 'mac', 'network_view', 'extattrs']
     _remap = {'ipv4addr': 'ip'}
@@ -759,8 +800,10 @@ class FixedAddressV4(FixedAddress):
 class FixedAddressV6(FixedAddress):
     """FixedAddress for IPv6"""
     _infoblox_type = 'ipv6fixedaddress'
-    _fields = ['ipv6addr', 'duid', 'network_view', 'extattrs', 'network']
-    _search_fields = ['ipv6addr', 'duid', 'network_view', 'network']
+    _fields = ['ipv6addr', 'duid', 'network_view', 'extattrs', 'network',
+               'comment']
+    _search_for_update_fields = ['ipv6addr', 'duid', 'network_view', 'network']
+    _all_searchable_fields = _search_for_update_fields
     _return_fields = ['ipv6addr', 'duid', 'network_view', 'extattrs']
     _shadow_fields = ['_ref', 'mac', 'ip']
     _remap = {'ipv6addr': 'ip'}
@@ -799,7 +842,8 @@ class ARecordBase(InfobloxObject):
 class ARecord(ARecordBase):
     _infoblox_type = 'record:a'
     _fields = ['ipv4addr', 'name', 'view', 'extattrs']
-    _search_fields = ['ipv4addr', 'view']
+    _search_for_update_fields = ['ipv4addr', 'view']
+    _all_searchable_fields = _search_for_update_fields + ['name']
     _shadow_fields = ['_ref']
     _remap = {'ip': 'ipv4addr'}
     _ip_version = 4
@@ -808,7 +852,8 @@ class ARecord(ARecordBase):
 class AAAARecord(ARecordBase):
     _infoblox_type = 'record:aaaa'
     _fields = ['ipv6addr', 'name', 'view', 'extattrs']
-    _search_fields = ['ipv6addr', 'view']
+    _search_for_update_fields = ['ipv6addr', 'view']
+    _all_searchable_fields = _search_for_update_fields + ['name']
     _shadow_fields = ['_ref']
     _remap = {'ip': 'ipv6addr'}
     _ip_version = 6
@@ -828,7 +873,8 @@ class PtrRecord(InfobloxObject):
 
 class PtrRecordV4(PtrRecord):
     _fields = ['view', 'ipv4addr', 'ptrdname', 'extattrs']
-    _search_fields = ['view', 'ipv4addr']
+    _search_for_update_fields = ['view', 'ipv4addr']
+    _all_searchable_fields = _search_for_update_fields + ['ptrdname']
     _shadow_fields = ['_ref']
     _remap = {'ip': 'ipv4addr'}
     _ip_version = 4
@@ -836,17 +882,35 @@ class PtrRecordV4(PtrRecord):
 
 class PtrRecordV6(PtrRecord):
     _fields = ['view', 'ipv6addr', 'ptrdname', 'extattrs']
-    _search_fields = ['view', 'ipv6addr']
+    _search_for_update_fields = ['view', 'ipv6addr']
+    _all_searchable_fields = _search_for_update_fields + ['ptrdname']
     _shadow_fields = ['_ref']
     _remap = {'ip': 'ipv6addr'}
     _ip_version = 6
+
+
+class SRVRecord(InfobloxObject):
+    _infoblox_type = 'record:srv'
+    _fields = ['name', 'port', 'priority', 'target', 'weight',
+               'aws_rte53_record_info', 'cloud_info', 'comment',
+               'creation_time', 'creator', 'ddns_principal',
+               'ddns_protected', 'disable', 'dns_name', 'dns_target',
+               'extattrs', 'forbid_reclamation', 'reclaimable',
+               'ttl', 'use_ttl', 'view', 'zone']
+    _search_for_update_fields = ['comment', 'creator', 'ddns_principal',
+                                 'name', 'port', 'priority', 'reclaimable',
+                                 'target', 'view', 'weight', 'zone']
+    _all_searchable_fields = _search_for_update_fields
+    _return_fields = ['name', 'port', 'priority', 'weight', 'target', 'view']
+    _shadow_fields = ['_ref']
 
 
 class NetworkView(InfobloxObject):
     _infoblox_type = 'networkview'
     _fields = ['name', 'extattrs']
     _return_fields = ['name', 'extattrs', 'is_default']
-    _search_fields = ['name']
+    _search_for_update_fields = ['name']
+    _all_searchable_fields = _search_for_update_fields
     _shadow_fields = ['_ref', 'is_default']
     _ip_version = 'any'
 
@@ -855,19 +919,21 @@ class DNSView(InfobloxObject):
     _infoblox_type = 'view'
     _fields = ['name', 'network_view', 'extattrs']
     _return_fields = ['name', 'network_view', 'extattrs']
-    _search_fields = ['name', 'network_view']
+    _search_for_update_fields = ['name', 'network_view']
+    _all_searchable_fields = _search_for_update_fields
     _shadow_fields = ['_ref', 'is_default']
     _ip_version = 'any'
 
 
 class DNSZone(InfobloxObject):
     _infoblox_type = 'zone_auth'
-    _fields = ['_ref', 'fqdn', 'view', 'extattrs', 'zone_format', 'ns_group',
+    _fields = ['fqdn', 'view', 'extattrs', 'zone_format',
                'prefix', 'grid_primary', 'grid_secondaries']
     _return_fields = ['fqdn', 'view', 'extattrs', 'zone_format', 'ns_group',
                       'prefix', 'grid_primary', 'grid_secondaries']
-    _search_fields = ['fqdn', 'view']
-    _shadow_fields = ['_ref']
+    _search_for_update_fields = ['fqdn', 'view', 'zone_format']
+    _all_searchable_fields = _search_for_update_fields
+    _shadow_fields = ['_ref', 'ns_group']
     _ip_version = 'any'
 
     @staticmethod
@@ -888,7 +954,8 @@ class Member(InfobloxObject):
                'config_addr_type', 'service_type_configuration']
     _return_fields = ['host_name', 'ipv6_setting', 'node_info',
                       'vip_setting', 'extattrs']
-    _search_fields = ['host_name', 'ipv4_address', 'ipv6_address']
+    _search_for_update_fields = ['host_name', 'ipv4_address', 'ipv6_address']
+    _all_searchable_fields = _search_for_update_fields
     _shadow_fields = ['_ref', 'ip', 'node_info']
     _ip_version = 'any'
     _remap = {'name': 'host_name'}
@@ -900,7 +967,8 @@ class EADefinition(InfobloxObject):
     _fields = ['comment', 'default_value', 'flags', 'list_values',
                'max', 'min', 'name', 'namespace', 'type',
                'allowed_object_types']
-    _search_fields = ['name']
+    _search_for_update_fields = ['name']
+    _all_searchable_fields = _search_for_update_fields
     _shadow_fields = ['_ref']
     _return_fields = ['comment', 'default_value', 'flags', 'list_values',
                       'max', 'min', 'name', 'namespace', 'type',
@@ -909,7 +977,9 @@ class EADefinition(InfobloxObject):
 
 class IPAddress(InfobloxObject):
     _fields = ['network_view', 'ip_address', 'objects', 'network', 'status']
-    _search_fields = ['network_view', 'ip_address', 'network', 'status']
+    _search_for_update_fields = ['network_view', 'ip_address',
+                                 'network', 'status']
+    _all_searchable_fields = _search_for_update_fields
     _shadow_fields = ['_ref']
     _return_fields = ['objects']
 
@@ -958,5 +1028,19 @@ class IPAllocation(object):
 class Tenant(InfobloxObject):
     _infoblox_type = 'grid:cloudapi:tenant'
     _fields = ['id', 'name', 'comment']
-    _search_fields = ['id']
+    _search_for_update_fields = ['id']
+    _all_searchable_fields = _search_for_update_fields
+    _shadow_fields = ['_ref']
+
+
+class CNAMERecord(InfobloxObject):
+    _infoblox_type = 'record:cname'
+    _fields = ['name', 'canonical', 'view', 'extattrs', 'comment',
+               'creator', 'ddns_principal', 'ddns_protected', 'disable',
+               'forbid_reclamation', 'ttl', 'use_ttl']
+    _search_for_update_fields = ['name', 'view']
+    _updateable_search_fields = ['name']
+    _all_searchable_fields = _search_for_update_fields + ['reclaimable',
+                                                          'zone']
+    _return_fields = ['canonical', 'name', 'view', 'extattrs']
     _shadow_fields = ['_ref']