+++ tests/test_connector.py
@@ -45,6 +45,7 @@ class TestInfobloxConnector(unittest.TestCase):
         opts.http_request_timeout = 10
         opts.max_retries = 3
         opts.max_results = None
+        opts.paging = False
         return opts
 
     def test_create_object(self):
@@ -123,7 +124,7 @@ class TestInfobloxConnector(unittest.TestCase):
             self.connector.get_object(objtype, payload, extattrs=extattrs)
             patched_get.assert_called_once_with(
                 'https://infoblox.example.org/wapi/'
-                'v1.1/network?*Subnet ID=fake_subnet_id&ip=0.0.0.0',
+                'v1.1/network?%2ASubnet+ID=fake_subnet_id&ip=0.0.0.0',
                 headers=self.connector.DEFAULT_HEADER,
                 timeout=self.default_opts.http_request_timeout,
             )
@@ -196,6 +197,28 @@ class TestInfobloxConnector(unittest.TestCase):
                 timeout=self.default_opts.http_request_timeout,
             )
 
+    def test_update_object_with_http_error(self):
+        ref = 'network'
+        payload = {'ip': '0.0.0.0'}
+
+        with patch.object(requests.Session, 'put',
+                          return_value=mock.Mock()) as patched_update:
+            patched_update.return_value.status_code = 400
+            patched_update.return_value.content = '{}'
+            self.assertRaises(exceptions.InfobloxCannotUpdateObject,
+                              self.connector.update_object, ref, payload)
+
+    def test_update_object_with_http_error_503(self):
+        ref = 'network'
+        payload = {'ip': '0.0.0.0'}
+
+        with patch.object(requests.Session, 'put',
+                          return_value=mock.Mock()) as patched_update:
+            patched_update.return_value.status_code = 503
+            patched_update.return_value.content = 'Temporary Unavailable'
+            self.assertRaises(exceptions.InfobloxGridTemporaryUnavailable,
+                              self.connector.update_object, ref, payload)
+
     def test_delete_object(self):
         ref = 'network'
         with patch.object(requests.Session, 'delete',
@@ -208,6 +231,24 @@ class TestInfobloxConnector(unittest.TestCase):
                 headers=self.connector.DEFAULT_HEADER,
                 timeout=self.default_opts.http_request_timeout,
             )
+
+    def test_delete_object_with_http_error(self):
+        ref = 'network'
+        with patch.object(requests.Session, 'delete',
+                          return_value=mock.Mock()) as patched_delete:
+            patched_delete.return_value.status_code = 400
+            patched_delete.return_value.content = '{}'
+            self.assertRaises(exceptions.InfobloxCannotDeleteObject,
+                              self.connector.delete_object, ref)
+
+    def test_delete_object_with_http_error_503(self):
+        ref = 'network'
+        with patch.object(requests.Session, 'delete',
+                          return_value=mock.Mock()) as patched_delete:
+            patched_delete.return_value.status_code = 503
+            patched_delete.return_value.content = 'Temporary Unavailable'
+            self.assertRaises(exceptions.InfobloxGridTemporaryUnavailable,
+                              self.connector.delete_object, ref)
 
     def test_construct_url_absolute_path_fails(self):
         pathes = ('/starts_with_slash', '', None)
@@ -222,7 +263,7 @@ class TestInfobloxConnector(unittest.TestCase):
                                             query_params=query_params,
                                             extattrs=ext_attrs)
         self.assertEqual('https://infoblox.example.org/wapi/v1.1/network?'
-                         '*Subnet ID=fake_subnet_id&some_option=some_value',
+                         '%2ASubnet+ID=fake_subnet_id&some_option=some_value',
                          url)
 
     def test_construct_url_with_force_proxy(self):
@@ -231,7 +272,7 @@ class TestInfobloxConnector(unittest.TestCase):
                                             extattrs=ext_attrs,
                                             force_proxy=True)
         self.assertEqual('https://infoblox.example.org/wapi/v1.1/network?'
-                         '*Subnet ID=fake_subnet_id&_proxy_search=GM',
+                         '%2ASubnet+ID=fake_subnet_id&_proxy_search=GM',
                          url)
 
     def test_get_object_with_proxy_flag(self):
@@ -243,7 +284,7 @@ class TestInfobloxConnector(unittest.TestCase):
 
         self.assertEqual(None, result)
         self.connector._construct_url.assert_called_with('network', {},
-                                                         None, True)
+                                                         None, force_proxy=True)
         self.connector._get_object.called_with('network',
                                                self.connector._construct_url)
 
@@ -255,7 +296,7 @@ class TestInfobloxConnector(unittest.TestCase):
         result = self.connector.get_object('network')
 
         self.assertEqual(None, result)
-        construct_calls = [mock.call('network', {}, None, False),
+        construct_calls = [mock.call('network', {}, None, force_proxy=False),
                            mock.call('network', {}, None, force_proxy=True)]
         self.connector._construct_url.assert_has_calls(construct_calls)
 
@@ -268,6 +309,127 @@ class TestInfobloxConnector(unittest.TestCase):
 
         url = 'http://some-url/'
         self.assertEqual(None, self.connector._get_object('network', url))
+
+    def test_get_object_with_pagination_with_no_result(self):
+         self.connector._get_object = mock.MagicMock(return_value=None)
+         result = self.connector.get_object('network', paging=True)
+         self.assertEqual(None, result)
+
+    def test_get_object_with_pagination_with_result(self):
+         self.connector._get_object = mock.MagicMock(
+                                          return_value={"result": ["data"]})
+         result = self.connector.get_object('network', paging=True)
+         self.assertEqual(["data"], result)
+
+    def test__handle_get_object_with_pagination_with_no_record(self):
+        query_params = {"_paging": 1,
+                        "_return_as_object": 1,
+                        "_max_results": 100}
+        self.connector._get_object = mock.MagicMock(return_value=None)
+        result = self.connector._handle_get_object("network", query_params,
+                                                   None, False)
+        self.assertEqual(None, result)
+
+    def test__handle_get_object_with_max_results_nigative(self):
+        query_params = {"_paging": 1,
+                        "_return_as_object": 1,
+                        "_max_results": -100}
+        self.connector._get_object = mock.MagicMock(return_value=None)
+        result = self.connector._handle_get_object("network", query_params,
+                                                   None, False)
+        self.assertEqual(None, result)
+
+    def test__handle_get_object_with_pagination_with_record(self):
+        query_params = {"_paging": 1,
+                        "_return_as_object": 1,
+                        "_max_results": 100}
+        self.connector._get_object = mock.MagicMock(
+                                         return_value={"result": ["data"]})
+        result = self.connector._handle_get_object("network", query_params,
+                                                   None, False)
+        self.assertEqual(["data"], result)
+
+    def _get_object(self, url, **opts):
+        resp = requests.Response
+        resp.status_code = 200
+        if "_page_id" in url:
+            resp.content = jsonutils.dumps({"result": [6,7,8,9,10]})
+        else:
+            resp.content = jsonutils.dumps(
+                               {"result": [1,2,3,4,5], "next_page_id": 1})
+        return resp
+
+    def test__handle_get_object_with_record_more_than_max_results_paging(self):
+        query_params = {"_paging": 1,
+                        "_return_as_object": 1,
+                        "_max_results": 5}
+        with patch.object(requests.Session, 'get') as patched_get:
+            patched_get.side_effect = self._get_object
+            result = self.connector._handle_get_object("network", query_params,
+                                                       None, False)
+        self.assertEqual([1,2,3,4,5,6,7,8,9,10], result)
+
+    def test__handle_get_object_without_pagination(self):
+        query_params = {"_max_results": 100}
+        self.connector._get_object = mock.MagicMock(return_value=None)
+        result = self.connector._handle_get_object("network", query_params,
+                                                   None, False)
+        self.assertEqual(None, result)
+
+    def test__handle_get_object_without_pagination_with_record(self):
+        query_params = {"_max_results": 100}
+        self.connector._get_object = mock.MagicMock(return_value=["data"])
+        result = self.connector._handle_get_object("network", query_params,
+                                                   None, False)
+        self.assertEqual(["data"], result)
+
+    def test_call_func(self):
+        objtype = 'network'
+        payload = {'ip': '0.0.0.0'}
+
+        with patch.object(requests.Session, 'post',
+                          return_value=mock.Mock()) as patched_call_func:
+            patched_call_func.return_value.status_code = 201
+            patched_call_func.return_value.content = '{}'
+            self.connector.call_func(objtype, "_ref", payload)
+            patched_call_func.assert_called_once_with(
+                'https://infoblox.example.org/wapi/v1.1/_ref?_function=network',
+                data=jsonutils.dumps(payload),
+                headers=self.connector.DEFAULT_HEADER,
+                timeout=self.default_opts.http_request_timeout,
+            )
+
+    def test_call_func_with_http_error(self):
+        objtype = 'network'
+        payload = {'ip': '0.0.0.0'}
+
+        with patch.object(requests.Session, 'post',
+                          return_value=mock.Mock()) as patched_call_func:
+            patched_call_func.return_value.status_code = 400
+            patched_call_func.return_value.content = '{}'
+            self.assertRaises(exceptions.InfobloxFuncException,
+                              self.connector.call_func, objtype, "_ref", payload)
+
+    def test_call_func_with_http_error_503(self):
+        objtype = 'network'
+        payload = {'ip': '0.0.0.0'}
+
+        with patch.object(requests.Session, 'post',
+                          return_value=mock.Mock()) as patched_call_func:
+            patched_call_func.return_value.status_code = 503
+            patched_call_func.return_value.content = 'Temporary Unavailable'
+            self.assertRaises(exceptions.InfobloxGridTemporaryUnavailable,
+                              self.connector.call_func, objtype, "_ref", payload)
+
+    def test__check_service_availability(self):
+        objtype = 'network'
+        payload = {'ip': '0.0.0.0'}
+        resp = requests.Response
+        resp.status_code = 503
+        resp.content = 'Temporary Unavailable'
+        self.assertRaises(exceptions.InfobloxGridTemporaryUnavailable,
+                          self.connector._check_service_availability, "delete",
+                          resp, '_ref')
 
 
 class TestInfobloxConnectorStaticMethods(unittest.TestCase):